Remove warning
