+++ src/Menus/History.menu.js
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
+import { View } from 'react-native';
 
 import { HistoryTabContainer } from '../Containers';
 import colors from '../Assets/Colors';
@@ -9,25 +9,23 @@ import I18n from '../Localization';
 const { primary } = colors;
 
 const History = () => {
-  const Stack = createStackNavigator();
+  const { Navigator, Screen } = createStackNavigator();
 
   return (
-    <NavigationContainer>
-      <Stack.Navigator>
-        <Stack.Screen
-          name="History"
-          component={HistoryTabContainer}
-          options={{
-            title: I18n.t('history'),
-            headerStyle: {
-              elevation: 0,
-              shadowOpacity: 0
-            },
-            headerTintColor: primary.black
-          }}
-        />
-      </Stack.Navigator>
-    </NavigationContainer>
+    <Navigator>
+      <Screen
+        name="History"
+        component={HistoryTabContainer}
+        options={{
+          title: I18n.t('history'),
+          headerStyle: {
+            elevation: 0,
+            shadowOpacity: 0
+          },
+          headerTintColor: primary.black
+        }}
+      />
+    </Navigator>
   );
 };
